requirements.txt
blinker==1.8.2
click==8.1.7
Flask==3.0.3
Flask-Cors==4.0.1
gunicorn==22.0.0
importlib_metadata==8.2.0
itsdangerous==2.2.0
Jinja2==3.1.4
lxml==5.2.2
MarkupSafe==2.1.5
packaging==24.1
PyPDF2==3.0.1
python-docx==1.1.2
python-dotenv==1.0.1
typing_extensions==4.12.2
Werkzeug==3.0.3
zipp==3.19.2
